ДЗ№8 8) Вынесли binding в отдельный класс ViewBindingFragment
