+++ app/src/main/java/com/example/materialdesign/view/ux/TutorialUXFragment.kt
@@ -14,25 +14,9 @@ import smartdevelop.ir.eram.showcaseviewlib.config.Gravity
 import smartdevelop.ir.eram.showcaseviewlib.listener.GuideListener
 
 
-class TutorialUXFragment : Fragment() {
-
-    private var _binding: FragmentUxTutorialBinding? = null
-    private val binding: FragmentUxTutorialBinding
-        get() = _binding!!
-
-
+class TutorialUXFragment : ViewBindingFragment<FragmentUxTutorialBinding>(FragmentUxTutorialBinding::inflate) {
     companion object {
         fun newInstance() = TutorialUXFragment()
-    }
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentUxTutorialBinding.inflate(inflater, container, false)
-        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -77,11 +61,6 @@ class TutorialUXFragment : Fragment() {
 
     }
 
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
 }
